pyproject.toml
[project]
name = "nix-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "brotli>=1.2.0",
    "pyyaml>=6.0.3",
    "textual>=7.4.0",
    "textual-dev>=1.8.0",
]
